Added state and getters
