+++ app/src/store/store.js
@@ -5,11 +5,14 @@ import Vuex from 'vuex'
 import VuexPersistence from 'vuex-persist'
 import config from '../config'
 import torus from '../torus'
-import { MAINNET, RPC } from '../utils/enums'
-import { formatCurrencyNumber, getRandomNumber, significantDigits, getEtherScanHashLink, broadcastChannelOptions } from '../utils/utils'
+import { RPC } from '../utils/enums'
+import { getRandomNumber, getEtherScanHashLink, broadcastChannelOptions } from '../utils/utils'
 import { post, get, patch } from '../utils/httpHelpers.js'
 import jwtDecode from 'jwt-decode'
 import { notifyUser } from '../utils/notifications'
+
+import initialState from './state'
+import getters from './getters'
 
 const web3Utils = torus.web3.utils
 
@@ -53,93 +56,13 @@ const vuexPersist = new VuexPersistence({
 
 var walletWindow
 
-const initialState = {
-  userInfo: {
-    email: '',
-    name: '',
-    profileImage: ''
-  },
-  idToken: '',
-  wallet: {}, // Account specific object
-  weiBalance: {}, // Account specific object
-  selectedAddress: '',
-  selectedCurrency: 'USD',
-  networkId: 0,
-  networkType: localStorage.getItem('torus_network_type') || MAINNET,
-  currencyData: {},
-  tokenData: {}, // Account specific object
-  tokenRates: {},
-  transactions: [],
-  unapprovedTypedMessages: {},
-  unapprovedPersonalMsgs: {},
-  unapprovedMsgs: {},
-  loginInProgress: false,
-  rpcDetails: JSON.parse(localStorage.getItem('torus_custom_rpc')) || {},
-  jwtToken: '',
-  pastTransactions: [],
-  isNewUser: false
-}
-
 var VuexStore = new Vuex.Store({
   plugins: [vuexPersist.plugin],
   state: {
     ...initialState
   },
   getters: {
-    unApprovedTransactions: state => {
-      const transactions = []
-      for (let id in state.transactions) {
-        if (state.transactions[id].status === 'unapproved') {
-          transactions.push(state.transactions[id])
-        }
-      }
-      return transactions
-    },
-    tokenBalances: state => {
-      let { weiBalance, tokenData, tokenRates, currencyData, selectedCurrency, networkType, selectedAddress } = state || {}
-      if (networkType !== MAINNET) {
-        tokenData = {}
-        tokenRates = {}
-      }
-      let currencyMultiplier = 1
-      if (selectedCurrency !== 'ETH') currencyMultiplier = currencyData[selectedCurrency.toLowerCase()] || 1
-      let full = [
-        {
-          balance: weiBalance[selectedAddress],
-          decimals: 18,
-          erc20: false,
-          logo: 'eth.svg',
-          name: 'Ethereum',
-          symbol: 'ETH',
-          tokenAddress: '0x'
-        }
-      ]
-      // because vue/babel is stupid
-      if (tokenData && tokenData[selectedAddress] && Object.keys(tokenData[selectedAddress]).length > 0) {
-        full = [...full, ...tokenData[selectedAddress]]
-      }
-      let totalPortfolioValue = 0
-      const finalBalancesArray = full.map(x => {
-        const computedBalance = parseFloat(torus.web3.utils.hexToNumberString(x.balance)) / 10 ** parseFloat(x.decimals) || 0
-        let tokenRateMultiplier = 1
-        if (x.tokenAddress !== '0x') tokenRateMultiplier = tokenRates[x.tokenAddress.toLowerCase()] || 0
-        const currencyRate = currencyMultiplier * tokenRateMultiplier
-        let currencyBalance = significantDigits(computedBalance * currencyRate || 0, false, 3)
-        totalPortfolioValue += currencyBalance
-        if (selectedCurrency !== 'ETH') currencyBalance = formatCurrencyNumber(currencyBalance)
-        return {
-          ...x,
-          id: x.symbol,
-          computedBalance: computedBalance,
-          formattedBalance: `${x.symbol} ${significantDigits(computedBalance || 0, false, 3)}`,
-          currencyBalance: `${selectedCurrency} ${currencyBalance}`,
-          currencyRateText: `1 ${x.symbol} = ${significantDigits(currencyRate || 0)} ${selectedCurrency}`
-        }
-      })
-      totalPortfolioValue = significantDigits(totalPortfolioValue, false, 3) || 0
-      if (selectedCurrency !== 'ETH') totalPortfolioValue = formatCurrencyNumber(totalPortfolioValue)
-      return { finalBalancesArray, totalPortfolioValue }
-    }
+    ...getters
   },
   mutations: {
     setUserInfo(state, userInfo) {
